+++ Magic/CHANGELOG.md
@@ -1,5 +1,7 @@
 # CHANGELOG
 
+# 10.8.2
+
 # 10.8.1
 
 ## Fixes